requirements.txt
fastapi
Pillow
qrcode
python-multipart
jinja2
